requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
python-dotenv==1.1.0
sqlparse==0.5.3
psycopg2-binary
supabase
Pillow